fix: Set Kafka permissions during install
The permissions for the meta.properties and bootstrap.checkpoint file are owned by the user root instead of Kafka.
Adding a task to ensure the permissions for the local log files for Kafka are set correctly. Added a variable for the
Kafka user, group and also the log directory.

Resolve: #50